--- a/src/main/java/org/apache/sysds/runtime/data/DenseBlockFP32.java
+++ b/src/main/java/org/apache/sysds/runtime/data/DenseBlockFP32.java
@@ -132,6 +132,13 @@ public class DenseBlockFP32 extends DenseBlockDRB
 		Arrays.fill(_data, fromIndex, toIndex, (float)v);
 	}
 
+	@Override 
+	public void fillRow(int r, double v){
+		int start = pos(r);
+		int end = start + getDim(1);
+		Arrays.fill(_data, start, end, (float)v);
+	}
+
 	@Override
 	protected void setInternal(int bix, int ix, double v) {
 		_data[ix] = (float)v;