--- a/src/main/java/org/apache/sysds/runtime/data/DenseBlockInt32.java
+++ b/src/main/java/org/apache/sysds/runtime/data/DenseBlockInt32.java
@@ -132,6 +132,13 @@ public class DenseBlockInt32 extends DenseBlockDRB
 		Arrays.fill(_data, fromIndex, toIndex, UtilFunctions.toInt(v));
 	}
 
+	@Override 
+	public void fillRow(int r, double v){
+		int start = pos(r);
+		int end = start + getDim(1);
+		Arrays.fill(_data, start, end, UtilFunctions.toInt(v));
+	}
+
 	@Override
 	protected void setInternal(int bix, int ix, double v) {
 		_data[ix] = UtilFunctions.toInt(v);