--- a/src/main/java/org/apache/sysds/runtime/data/DenseBlockLBool.java
+++ b/src/main/java/org/apache/sysds/runtime/data/DenseBlockLBool.java
@@ -150,6 +150,13 @@ public class DenseBlockLBool extends DenseBlockLDRB
 		_blocks[bix].set(fromIndex, toIndex, v != 0);
 	}
 
+	@Override 
+	public void fillRow(int r, double v){
+		int start = pos(r);
+		int end = start + getDim(1);
+		_blocks[index(r)].set(start, end, v != 0);
+	}
+
 	@Override
 	public DenseBlock set(String s) {
 		boolean b = Boolean.parseBoolean(s);