--- a/src/main/java/org/apache/sysds/runtime/data/DenseBlockLFP64DEDUP.java
+++ b/src/main/java/org/apache/sysds/runtime/data/DenseBlockLFP64DEDUP.java
@@ -178,6 +178,11 @@ public class DenseBlockLFP64DEDUP extends DenseBlockLDRB{
 		throw new NotImplementedException();
 	}
 
+	@Override 
+	public void fillRow(int r, double v){
+		throw new NotImplementedException();
+	}
+
 	@Override
 	protected void setInternal(int bix, int ix, double v) {
 		throw new NotImplementedException();