--- a/src/main/java/org/apache/sysds/runtime/data/DenseBlockLInt64.java
+++ b/src/main/java/org/apache/sysds/runtime/data/DenseBlockLInt64.java
@@ -109,6 +109,13 @@ public class DenseBlockLInt64 extends DenseBlockLDRB
 		Arrays.fill(_blocks[bix], fromIndex, toIndex, UtilFunctions.toLong(v));
 	}
 
+	@Override 
+	public void fillRow(int r, double v){
+		int start = pos(r);
+		int end = start + getDim(1);
+		Arrays.fill(_blocks[index(r)], start, end, UtilFunctions.toLong(v));
+	}
+
 	@Override
 	public DenseBlock set(int r, int c, double v) {
 		_blocks[index(r)][pos(r, c)] = UtilFunctions.toLong(v);