--- a/src/main/java/org/apache/sysds/runtime/matrix/data/LibMatrixReplace.java
+++ b/src/main/java/org/apache/sysds/runtime/matrix/data/LibMatrixReplace.java
@@ -151,13 +151,14 @@ public class LibMatrixReplace {
 		DenseBlock c = ret.getDenseBlock();
 
 		// initialize with replacement (since all 0 values, see SPARSITY_TURN_POINT)
-		c.reset(in.rlen, in.clen, replacement);
+		// c.reset(in.rlen, in.clen, replacement);
 
 		if(a == null)// check for empty matrix
 			return ((long) in.rlen) * in.clen;
 
 		// overwrite with existing values (via scatter)
 		for(int i = 0; i < in.rlen; i++) {
+			c.fillRow(i, replacement);
 			if(!a.isEmpty(i)) {
 				int apos = a.pos(i);
 				int cpos = c.pos(i);