This PR separate the replace logic from the internals of the MatrixBlock, adding a LibReplace class. 

The change follows the style of adding lib packages, and it improves the replace performance by .... (see next comment.)