@Baunsgaard 2024-07-04
The main benefit in this PR is when we perform replacement on a sparse matrix, and the output is dense. 

on a 10k x 10k 0.1 sparse matrix

After
```
java -jar target/systemds-3.3.0-SNAPSHOT-perf.jar 17 10000 10000 0.1 16
Profiling started
                        replaceZero,  132.883+-  9.335 ms,           
                         replaceOne,   59.612+-  3.295 ms,           
                         replaceNaN,   10.651+-  0.470 ms,
```

Before
```
java -jar target/systemds-3.3.0-SNAPSHOT-perf.jar 17 10000 10000 0.1 1
Profiling started
                        replaceZero,  228.727+- 11.965 ms,           
                         replaceOne,  163.212+-  4.993 ms,           
                         replaceNaN,   10.602+-  0.437 ms, 
```